Others

+++ app/Exports/Sheets/Region.php
@@ -148,7 +148,7 @@ class Region implements FromCollection, ShouldQueue, ShouldAutoSize, WithStyles,
             $price = str_replace(',', '.', $tablet_data['price']);
             $tablet_data['all_sales_price'] = $price*intval($tablet_data['all_sales']);
             $this->tablets[1]['all_sales'] = $this->tablets[1]['all_sales']+$tablet_data['all_sales'];
-            $this->tablets[1]['all_sales_price'] = $this->tablets[1]['all_sales_price']+($price*intval($tablet_data['all_sales']));
+            $this->tablets[1]['all_sales_price'] = $this->tablets[1]['all_sales_price']+($price*floatval($tablet_data['all_sales']));
             $this->tablets[] = $tablet_data;
             for ($i = 1; $i<=12; $i++){
                 $this->tablets[1][$i] += $tablet_data[$i];
@@ -178,11 +178,11 @@ class Region implements FromCollection, ShouldQueue, ShouldAutoSize, WithStyles,
                 if ($file->uploaded_date){
                     $price = str_replace(',', '.', $data['price']);
                     $data[Carbon::make($file->uploaded_date)->month] += $tablet->sales_qty;
-                    $data[Carbon::make($file->uploaded_date)->month+20] += intval($tablet->sales_qty)*intval($price);
+                    $data[Carbon::make($file->uploaded_date)->month+20] += floatval($tablet->sales_qty)*floatval($price);
                 }else{
                     $price = str_replace(',', '.', $data['price']);
                     $data[Carbon::now()->month] += $tablet->sales_qty;
-                    $data[Carbon::now()->month+20] += intval($tablet->sales_qty)*intval($price);
+                    $data[Carbon::now()->month+20] += floatval($tablet->sales_qty)*floatval($price);
                 }
             }
         }
